Fixat listan så nu kan man addera foods till en user och det ritar ut

+++ Code/src/model/BarChartEmissions.java
@@ -16,7 +16,7 @@ public class BarChartEmissions {
         barChart.getData().add(this.emissionSeries);
     }
 
-    public void addToChart(Date date, int emission){
+    public void addToChart(Date date, double emission){
         emissionSeries.getData().add(new XYChart.Data("" +date, emission));
 
     }
